[refactor] ProjectDetail - UserList 컴포넌트 분리, 이름 변경
- UserList Component 분리
  UserList의 헤더와 툴바를 파일로 분리

- Component 이름 변경
   기존 Material ui의 이름 삭제후,component에 맞는 이름으로 변경
