+++ src/components/ProjectDetail/ProjectDetailUserListToolbar.tsx
@@ -1,0 +1,67 @@
+import React from 'react';
+import clsx from 'clsx';
+import { createStyles, lighten, makeStyles, Theme } from '@material-ui/core/styles';
+import IconButton from '@material-ui/core/IconButton';
+import Tooltip from '@material-ui/core/Tooltip';
+import DeleteIcon from '@material-ui/icons/Delete';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+
+interface ProjectDetailUserListToolbarProps {
+  numSelected: number;
+  selectedUsers: number[];
+  deleteUsers: (selectedUids: number[]) => void;
+}
+
+const useToolbarStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      paddingLeft: theme.spacing(2),
+      paddingRight: theme.spacing(1),
+    },
+    highlight:
+      theme.palette.type === 'light'
+        ? {
+            color: theme.palette.secondary.main,
+            backgroundColor: lighten(theme.palette.secondary.light, 0.85),
+          }
+        : {
+            color: theme.palette.text.primary,
+            backgroundColor: theme.palette.secondary.dark,
+          },
+    title: {
+      flex: '1 1 100%',
+    },
+  }),
+);
+
+export default function ProjectDetailUserListToolbar(
+  props: ProjectDetailUserListToolbarProps,
+): React.ReactElement {
+  const classes = useToolbarStyles();
+  const { numSelected, deleteUsers, selectedUsers } = props;
+  return (
+    <Toolbar
+      className={clsx(classes.root, {
+        [classes.highlight]: numSelected > 0,
+      })}
+    >
+      {numSelected > 0 ? (
+        <Typography className={classes.title} color="inherit" variant="subtitle1" component="div">
+          {numSelected} selected
+        </Typography>
+      ) : (
+        <Typography className={classes.title} variant="h6" id="tableTitle" component="div">
+          Users
+        </Typography>
+      )}
+      {numSelected > 0 && (
+        <Tooltip title="Delete" onClick={() => deleteUsers(selectedUsers)}>
+          <IconButton aria-label="delete">
+            <DeleteIcon />
+          </IconButton>
+        </Tooltip>
+      )}
+    </Toolbar>
+  );
+}
